MDL-29663 forum: Forum read tracking On should force user read tracking.
If a forum is set to Read Tracking is On, it should be forced on the
user, just like Off is.

Logic change in forum_tp_can_track_forums().

Also required refactoring of forum_cm_info_view() which is somewhat
less efficient, as we need to do the db load of each forum, and
do more computations of each call. The privious version assume
that it would stay the same for all forums.