MDL-47003 atto: Improve paste behaviour and various cleanups
Background: The old code cleaned all content very heavily and also
failed to clean content in the editor div, which resulted in
undesirable code being being left in the div for editing, and also
stripping too much out of allowed code.

The solution is multi-fold. We now separate the code cleaning into
two different types of clean - one on standard clean event (update
textarea or loading editor from texteach) and one that only fires on
paste events (ideally only on html paste events).

When the paste event it fired, pasteCleanup() is called with the event.
If it has a clipboard we can access, we stop the event, take the
clipboard data, and run it through the extended scrubber and then
insert it into the editor. If we can't access the clipboard, register
fallbackPasteCleanup() to fire, which will clean the entire editor div.

We now allow html comments and unicode characters, and we try to only
strip Office like garbage on html paste events. We also now remove
entire style blocks, not just the open and close tags.