MDL-51833 accesslib: Improve event monitor lookup of capability
Event monitor needs to know if a user has the capability to view the
monitor in any course so it can insert a preference item in the user
menu.

Due to a lack of caching in the nav lib, this happens on every page
load, so we want to do our best to streamline the check.